# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dental Story API",
    "description": "API for the Dental Story dental clinic website. Provides endpoints for appointments, contact forms, and admin analytics.",
    "version": "1.0.0",
    "contact": {
      "name": "Dental Story Support",
      "email": "support@dentalstory.ua"
    }
  },
  "servers": [
    {
      "url": "https://dentalstory.ua",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/api/appointments": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Get appointments",
        "description": "Retrieve a list of appointments. Requires authentication for admin access.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Appointment" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Appointments"],
        "summary": "Create appointment",
        "description": "Book a new dental appointment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AppointmentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Appointment" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/appointments/slots": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Get available slots",
        "description": "Retrieve available appointment slots for a specific date",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-03-15"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "required": false,
            "description": "Optional doctor ID to filter slots",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available slots retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TimeSlot" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Appointment": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "patientName": { "type": "string" },
          "patientPhone": { "type": "string" },
          "patientEmail": { "type": "string", "format": "email" },
          "service": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "time": { "type": "string" },
          "doctorId": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled", "completed"]
          },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AppointmentRequest": {
        "type": "object",
        "required": ["patientName", "patientPhone", "service", "date", "time"],
        "properties": {
          "patientName": { "type": "string", "minLength": 2, "maxLength": 100 },
          "patientPhone": { "type": "string", "pattern": "^\\+380\\d{9}$" },
          "patientEmail": { "type": "string", "format": "email" },
          "service": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "time": { "type": "string" },
          "doctorId": { "type": "string" },
          "notes": { "type": "string", "maxLength": 500 },
          "turnstileToken": {
            "type": "string",
            "description": "Cloudflare Turnstile verification token"
          }
        }
      },
...
      "TimeSlot": {
        "type": "object",
        "properties": {
          "time": { "type": "string", "example": "09:00" },
          "available": { "type": "boolean" },
          "doctorId": { "type": "string" },
          "doctorName": { "type": "string" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin authentication token"
      }
    }
  }
}